nuke ToCBuilderZA, update base, move tests over
