+++ indigo_za/apps.py
@@ -10,5 +10,4 @@ class IndigoZAAppConfig(AppConfig):
         import indigo_za.publications  # noqa
         import indigo_za.refs  # noqa
         import indigo_za.terms  # noqa
-        import indigo_za.toc  # noqa
         import indigo_za.work_detail  # noqa
